+++ lib/auth/index.js
@@ -5,7 +5,7 @@ const LocalStrategy = require('passport-local').Strategy;
 const BasicStrategy = require('passport-http').BasicStrategy;
 const ClientPasswordStrategy = require('passport-oauth2-client-password').Strategy;
 const BearerStrategy = require('passport-http-bearer').Strategy;
-const db = require('../db');
+const db = require('../db/index');
 
 /**
  * LocalStrategy
