+++ lib/routes/oauth2.js
@@ -3,8 +3,8 @@
 const oauth2orize = require('oauth2orize');
 const passport = require('passport');
 const login = require('connect-ensure-login');
-const db = require('../db');
-const utils = require('../utils');
+const db = require('../db/index');
+const utils = require('../utils/index');
 
 // Create OAuth 2.0 server
 const server = oauth2orize.createServer();
